+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - `/users/updateEmail` endpoint added as `Iterable::Users#update_email`
 - `/users/{email}` delete endpoint added as `Iterable::Users#delete`
 - `/users/byUserId/{userId}` delete endpoint added as `Iterable::Users#delete_by_id`
+- `/users/registerBrowserToken` endpoint added as `Iterable::Users#register_browser_token`
 
 ### Workflows